# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/go-api/appbuilder": {
      "post": {
        "description": "Realiza uma consulta de carros por um request de uma ou mais linhas",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "AppBuilder"
        ],
        "summary": "Search",
        "parameters": [
          {
            "description": "Linhas de pesquisa",
            "name": "search",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appbuilder.appBuilderSearchRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/appbuilder.appBuilderResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/auth/aboutme": {
      "get": {
        "description": "Retorna informações do usuário.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "aboutMe",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/auth.aboutMeResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/auth/login": {
      "post": {
        "description": "Realiza um login de usuário",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "parameters": [
          {
            "description": "Credenciais",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth.loginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/auth.loginResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/go-api/auth/register": {
      "post": {
        "description": "Cria um novo usuário.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "parameters": [
          {
            "description": "Dados do usuário",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth.registerRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/auth.registerResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/export/apps": {
      "post": {
        "description": "Baixa as aplicações informadas no formato de csv",
        "produces": [
          "text/csv"
        ],
        "tags": [
          "Download"
        ],
        "summary": "Applications",
        "parameters": [
          {
            "description": "Ids dos veiculos",
            "name": "id",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/download.appDownloadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo CSV",
            "schema": {
              "type": "file"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/export/images": {
      "get": {
        "description": "Baixa todas as imagens de um produto",
        "produces": [
          "image/png",
          "image/jpeg"
        ],
        "tags": [
          "Download"
        ],
        "summary": "Images",
        "parameters": [
          {
            "type": "string",
            "description": "Nome base dos arquivos",
            "name": "fileName",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "ID do produto",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "Indice da imagem",
            "name": "index",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "appbuilder.appBuilderResponse": {
      "type": "object",
      "properties": {
        "carIds": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/appbuilder.appBuilderSearchResponse"
          }
        },
        "total": {
          "type": "integer"
        }
      }
    },
    "appbuilder.appBuilderSearchRequest": {
      "type": "object",
      "properties": {
        "search": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "Fox Plus 1.6 8V 2009"
          ]
        }
      }
    },
    "appbuilder.appBuilderSearchResponse": {
      "type": "object",
      "properties": {
        "cars": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/appbuilder.car"
          }
        },
        "search": {
          "type": "string",
          "example": "Fox Plus 1.6 8V 2009"
        }
      }
    },
    "appbuilder.car": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer",
          "example": 20976
        },
        "configMotor": {
          "type": "string",
          "example": "1.6 8V"
        },
        "filterData": {
          "$ref": "#/definitions/global.JsonMap"
        },
        "make": {
          "type": "string",
          "example": "VOLKSWAGEN"
        },
        "model": {
          "type": "string",
          "example": "Fox"
        },
        "version": {
          "type": "string",
          "example": "Plus 1.6 2009"
        },
        "year": {
          "type": "integer",
          "example": 2009
        }
      }
    },
    "auth.aboutMeResponse": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string"
        },
        "user_id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "auth.loginRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "example": "123"
        },
        "username": {
          "type": "string",
          "example": "user"
        }
      }
    },
    "auth.loginResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Login realizado com sucesso"
        },
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
...
    "auth.registerRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "example": "123"
        },
        "username": {
          "type": "string",
          "example": "user"
        }
      }
    },
    "auth.registerResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "User created"
        }
      }
    },
    "download.appDownloadRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "integer",
          "example": 1
        },
        "vehicleId": {
          "type": "array",
          "items": {
            "type": "integer"
          },
          "example": [
            1
          ]
        }
      }
    },
    "global.JsonMap": {
      "type": "object",
      "additionalProperties": {}
    },
    "httpcustom.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      }
    },
...
  }
}